requirements.txt
python-dotenv==1.0.*
python-telegram-bot==13.14
google-cloud==0.34.0
google-cloud-api-keys==0.5.3
google-cloud-dialogflow==2.23.3
google-cloud-speech==2.21.0